+++ src/main/java/com/wenxin2/marioverse/items/DashMushroomItem.java
@@ -48,7 +48,7 @@ public class DashMushroomItem extends Item {
             for (int lineAmt = 1; lineAmt <= tooltipLineAmt; lineAmt++) {
                 MutableComponent abilityText = Component.translatable(this.getDescriptionId() + ".tooltip.line" + lineAmt);
 
-                if (stack.is(ItemRegistry.DASH_MUSHROOM) && lineAmt == 3)
+                if (stack.is(ItemRegistry.DASH_MUSHROOM) && lineAmt == 5)
                     abilityText = abilityText.append(Component.translatable(this.getDescriptionId() + ".tooltip.line" + lineAmt + ".hearts",
                             ConfigRegistry.DASH_MUSHROOM_HEALTH_HEALED.get().floatValue()).withStyle(ChatFormatting.RED));