+++ src/main/java/com/wenxin2/marioverse/registries/BlockRegistry.java
@@ -1442,13 +1442,13 @@ public class BlockRegistry {
                                 .mapColor(color.getName().equals(DyeColor.WHITE.getName()) ? MapColor.TERRACOTTA_WHITE : color.getMapColor())))));
 
         CLASSIC_CHECKPOINT_FLAG = registerNoItemBlock("classic_checkpoint_flag",
-                () -> new CheckpointFlagBlock(4, null, BlockBehaviour.Properties.of().mapColor(MapColor.GOLD)
+                () -> new CheckpointFlagBlock(3, null, BlockBehaviour.Properties.of().mapColor(MapColor.GOLD)
                         .sound(SoundType.NETHERITE_BLOCK).isViewBlocking(BlockRegistry::never)
                         .strength(2.5F, 3.0F).requiresCorrectToolForDrops().noCollission()));
 
         Arrays.stream(DyeColor.values()).forEach(color ->
                 CHECKPOINT_FLAGS.put(color, registerNoItemBlock(color.getName() + "_checkpoint_flag",
-                        () -> new CheckpointFlagBlock(4, color, BlockBehaviour.Properties.of().mapColor(MapColor.GOLD)
+                        () -> new CheckpointFlagBlock(3, color, BlockBehaviour.Properties.of().mapColor(MapColor.GOLD)
                                 .sound(SoundType.NETHERITE_BLOCK).isViewBlocking(BlockRegistry::never)
                                 .strength(2.5F, 3.0F).requiresCorrectToolForDrops().noCollission()))));